requirements.txt
flask==2.2.3
notion-client==1.0.0
python-dotenv==1.0.0
requests==2.28.1
boto3==1.28.28